#14 戻るボタンのコンポーネント化
