+++ pages/_app.js
@@ -1,8 +1,9 @@
 import { ChakraProvider } from "@chakra-ui/react";
 import { RecoilRoot } from "recoil";
 import "../styles/globals.css";
+import theme from "../theme/theme";
 import { AuthProvider } from "./context/AuthContext";
-import theme from "./theme/theme";
+
 
 function MyApp({ Component, pageProps }) {
   return (
