+++ pages/editPosts.js
@@ -12,38 +12,21 @@ import {
   Tr,
   Box,
   Button,
-  Flex,
-  Heading,
-  Drawer,
-  DrawerOverlay,
-  DrawerContent,
-  DrawerCloseButton,
-  DrawerHeader,
-  DrawerBody,
-  Avatar,
-  DrawerFooter,
-  useDisclosure,
 } from "@chakra-ui/react";
 
 import Router from "next/router";
 import { postsState } from "./atoms/atom";
 import Head from "next/head";
-import {
-  AddIcon,
-  ArrowLeftIcon,
-  CheckIcon,
-  EditIcon,
-  HamburgerIcon,
-} from "@chakra-ui/icons";
+import { AddIcon, ArrowLeftIcon } from "@chakra-ui/icons";
 import { useRef } from "react";
+import Header from "../src/components/Header";
+import BackButton from "../src/components/atoms/BackButton";
 const handler = (path) => {
   Router.push(path);
 };
 
 const Edit = () => {
   const posts = useRecoilValue(postsState);
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const btnRef = useRef();
 
   return (
     <>
@@ -60,70 +43,7 @@ const Edit = () => {
         m="0"
         padding="1rem"
       >
-        <Heading pt="2" pb="2" pl="5" bg="#C5D8D1">
-          <Flex alignItems="center">
-            {/* エラー文 404 (Not Found) */}
-            <HamburgerIcon ref={btnRef} onClick={onOpen} mr="24px" />
-
-            <Drawer
-              isOpen={isOpen}
-              placement="left"
-              onClose={onClose}
-              finalFocusRef={btnRef}
-            >
-              <DrawerOverlay />
-              <DrawerContent>
-                <DrawerCloseButton />
-                <DrawerHeader>Your account</DrawerHeader>
-
-                <DrawerBody>
-                  <Flex mb="24px" alignItems="center">
-                    <Avatar src="https://bit.ly/broken-link" mr="8px" />
-                    <Text>リヴァイ</Text>
-                  </Flex>
-
-                  <Flex
-                    alignItems="center"
-                    cursor="pointer"
-                    onClick={() => handler("/posting")}
-                    _hover={{ background: "#8FBFE0", borderRadius: "20px" }}
-                  >
-                    <AddIcon mr="8px" />
-                    <Text lineHeight="48px">質問を作成</Text>
-                  </Flex>
-
-                  <Flex
-                    alignItems="center"
-                    cursor="pointer"
-                    mb="16px"
-                    _hover={{ background: "#8FBFE0", borderRadius: "20px" }}
-                    onClick={() => handler("/editPosts")}
-                  >
-                    <EditIcon mr="8px" />
-                    <Text lineHeight="48px">作成した質問を編集</Text>
-                  </Flex>
-                  <Box textAlign="center">
-                    <Button
-                      mt="40px"
-                      colorScheme="red"
-                      onClick={() => handler("/")}
-                    >
-                      Log Out
-                    </Button>
-                  </Box>
-                </DrawerBody>
-
-                <DrawerFooter>
-                  <Button variant="outline" mr={3} onClick={onClose}>
-                    Cancel
-                  </Button>
-                </DrawerFooter>
-              </DrawerContent>
-            </Drawer>
-            <Text>みんなの質問</Text>
-          </Flex>
-        </Heading>
-
+        <Header />
         <Container minH="calc(100% - 64px)" maxW="100%" bg="white" padding="5">
           <Box mt="10" mr="5" textAlign="right">
             <Button
@@ -180,15 +100,7 @@ const Edit = () => {
             </Table>
           </Container>
           <Box pos="absolute" bottom="8" right="0">
-            <Button
-              background="#F4D1AE"
-              _hover={{ opacity: "0.8" }}
-              onClick={() => handler("/user")}
-             mr={8}
-            >
-              <ArrowLeftIcon mr="2" />
-              戻る
-            </Button>
+            <BackButton />
           </Box>
         </Container>
       </Container>
