+++ pages/user.js
@@ -23,6 +23,7 @@ import Router, { useRouter } from "next/router";
 import { auth } from "./firebase";
 import { useRecoilState } from "recoil";
 import { postsState } from "./atoms/atom";
+import Header from "../src/components/Header";
 const handler = (path) => {
   Router.push(path);
 };
@@ -36,7 +37,6 @@ const user = () => {
   };
   const { isOpen, onOpen, onClose } = useDisclosure();
   const btnRef = useRef();
-
 
   return (
     <>
@@ -53,69 +53,7 @@ const user = () => {
         m="0"
         padding="1rem"
       >
-        <Heading pt="2" pb="2" pl="5" bg="#C5D8D1">
-          <Flex alignItems="center">
-            {/* エラー文 404 (Not Found) */}
-            <HamburgerIcon ref={btnRef} onClick={onOpen} mr="24px" />
-
-            <Drawer
-              isOpen={isOpen}
-              placement="left"
-              onClose={onClose}
-              finalFocusRef={btnRef}
-            >
-              <DrawerOverlay />
-              <DrawerContent>
-                <DrawerCloseButton />
-                <DrawerHeader>Your account</DrawerHeader>
-
-                <DrawerBody>
-                  <Flex mb="24px" alignItems="center">
-                    <Avatar src="https://bit.ly/broken-link" mr="8px" />
-                    <Text>リヴァイ</Text>
-                  </Flex>
-
-                  <Flex
-                    alignItems="center"
-                    cursor="pointer"
-                    onClick={() => handler("/posting")}
-                    _hover={{ background: "#8FBFE0", borderRadius: "20px" }}
-                  >
-                    <AddIcon mr="8px" />
-                    <Text lineHeight="48px">質問を作成</Text>
-                  </Flex>
-
-                  <Flex
-                    alignItems="center"
-                    cursor="pointer"
-                    mb="16px"
-                    _hover={{ background: "#8FBFE0", borderRadius: "20px" }}
-                    onClick={() => handler("/editPosts")}
-                  >
-                    <EditIcon mr="8px" />
-                    <Text lineHeight="48px">作成した質問を編集</Text>
-                  </Flex>
-                  <Box textAlign="center">
-                    <Button
-                      mt="40px"
-                      colorScheme="red"
-                      onClick={handleLogout}
-                    >
-                      Log Out
-                    </Button>
-                  </Box>
-                </DrawerBody>
-
-                <DrawerFooter>
-                  <Button variant="outline" mr={3} onClick={onClose}>
-                    Cancel
-                  </Button>
-                </DrawerFooter>
-              </DrawerContent>
-            </Drawer>
-            <Text>みんなの質問</Text>
-          </Flex>
-        </Heading>
+        <Header />
 
         <Container minH="calc(100% - 64px)" maxW="100%" bg="white" padding="5">
           <Box mt="10" mr="5" textAlign="right">
