+++ pages/user/[id]/answer.jsx
@@ -1,40 +1,25 @@
 import React, { useRef, useState } from "react";
-import {
-  AddIcon,
-  ArrowLeftIcon,
-  EditIcon,
-  HamburgerIcon,
-} from "@chakra-ui/icons";
+import { ArrowLeftIcon } from "@chakra-ui/icons";
 import {
   Box,
   Button,
   Container,
   Divider,
-  Drawer,
-  DrawerBody,
-  DrawerCloseButton,
-  DrawerContent,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerOverlay,
   Flex,
   FormControl,
   FormLabel,
-  Heading,
-  Input,
   Radio,
   RadioGroup,
   Spacer,
   Stack,
   Text,
-  useDisclosure,
-  useToast,
 } from "@chakra-ui/react";
-import { Avatar } from "@chakra-ui/react";
+
 import Head from "next/head";
 import Router, { useRouter } from "next/router";
-import { useRecoilState, useRecoilValue } from "recoil";
+import { useRecoilState } from "recoil";
 import { postsState } from "../../atoms/atom";
+import Header from "../../../src/components/Header";
 
 const handler = (path) => {
   Router.push(path);
@@ -45,14 +30,9 @@ const Answer = () => {
   const [value, setValue] = useState();
   // const [yesCount, setYesCount] = useState(Number);
   // const [noCount, setNoCount] = useState(Number);
-  const[ count, setCount ] = useState(0);
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const btnRef = useRef();
-  const router = useRouter();
+  // const [count, setCount] = useState(0);
 
- const incremant = () => {
-  setCount(count + 1)
- }
+  const router = useRouter();
 
   const post = posts.filter((post) => {
     return post.id === Number(router.query.id);
@@ -96,68 +76,7 @@ const Answer = () => {
         m="0"
         padding="1rem"
       >
-        <Heading pt="2" pb="2" pl="5" bg="#C5D8D1">
-          <Flex alignItems="center">
-            {/* エラー文 404 (Not Found) */}
-            <HamburgerIcon ref={btnRef} onClick={onOpen} mr="24px" />
-
-            <Drawer
-              isOpen={isOpen}
-              placement="left"
-              onClose={onClose}
-              finalFocusRef={btnRef}
-            >
-              <DrawerOverlay />
-              <DrawerContent>
-                <DrawerCloseButton />
-                <DrawerHeader>Your account</DrawerHeader>
-
-                <DrawerBody>
-                  <Flex mb="24px" alignItems="center">
-                    <Avatar src="https://bit.ly/broken-link" mr="8px" />
-                    <Text>リヴァイ</Text>
-                  </Flex>
-
-                  <Flex
-                    alignItems="center"
-                    cursor="pointer"
-                    onClick={() => handler("/posting")}
-                    _hover={{ background: "#8FBFE0", borderRadius: "20px" }}
-                  >
-                    <AddIcon mr="8px" />
-                    <Text lineHeight="48px">質問を作成</Text>
-                  </Flex>
-
-                  <Flex
-                    alignItems="center"
-                    mb="16px"
-                    _hover={{ background: "#8FBFE0", borderRadius: "20px" }}
-                  >
-                    <EditIcon mr="8px" />
-                    <Text lineHeight="48px">作成した質問を編集</Text>
-                  </Flex>
-                  <Box textAlign="center">
-                    <Button
-                      mt="40px"
-                      colorScheme="red"
-                      onClick={() => handler("/")}
-                    >
-                      Log Out
-                    </Button>
-                  </Box>
-                </DrawerBody>
-
-                <DrawerFooter>
-                  <Button variant="outline" mr={3} onClick={onClose}>
-                    Cancel
-                  </Button>
-                </DrawerFooter>
-              </DrawerContent>
-            </Drawer>
-            <Text>みんなの質問</Text>
-          </Flex>
-        </Heading>
-
+        <Header />
         <Container minH="calc(100% - 64px)" maxW="100%" bg="white" padding="5">
           <form>
             <Container py={["20px", "60px"]} maxW="container.lg">
